requirements.txt
streamlit
plotly.express
pandas
st-gsheets-connection
pandasql
scipy.stats
